+++ spconv/build.py
@@ -28,6 +28,8 @@ if project_is_installed(PACKAGE_NAME) and project_is_editable(
     from cumm.conv.main import ConvMainUnitTest
 
     from spconv.csrc.sparse.all import SpconvOps
+    from spconv.csrc.utils import BoxOps
+
     cu = GemmMainUnitTest(SHUFFLE_SIMT_PARAMS + SHUFFLE_VOLTA_PARAMS +
                           SHUFFLE_TURING_PARAMS)
     cu.namespace = "cumm.gemm.main"
@@ -38,7 +40,7 @@ if project_is_installed(PACKAGE_NAME) and project_is_editable(
     if InWindows:
         # windows have command line limit, so we use objects_folder to reduce command size.
         objects_folder = "objects"
-    pccm.builder.build_pybind([cu, convcu, SpconvOps()],
+    pccm.builder.build_pybind([cu, convcu, SpconvOps(), BoxOps()],
                               PACKAGE_ROOT / "core_cc",
                               namespace_root=PACKAGE_ROOT,
                               objects_folder=objects_folder,